pyproject.toml
[project]
    dependencies=["astropy>=7.2.0", "pydantic>=2.12.5"]
    description="Solar eclipse calculations using NASA Besselian elements"
    name="eclipse"
    readme="README.md"
    requires-python=">=3.12"
    version="0.1.0"

    [project.optional-dependencies]
        dev=["astronomy-engine>=2.1.19", "hypothesis>=6.150.1", "pytest>=9.0.2"]

    [project.scripts]
        eclipse="eclipse.cli:main"

[tool.pytest.ini_options]
    pythonpath="."

[tool.hatch.build.targets.wheel]
    packages=["eclipse"]

    [tool.hatch.build.targets.wheel.force-include]
        "eclipse/data"="eclipse/data"
